requirements.txt
asgiref
cachetools
certifi
charset-normalizer
Django
django-filter
djangorestframework
djangorestframework-simplejwt
drf-yasg
Faker
google-auth
google-cloud-secret-manager
gunicorn
idna
inflection
mysqlclient
packaging
pandas
pillow
pyasn1
pyasn1_modules
pycountry
PyJWT
python-dateutil
python-decouple
python-dotenv
pytz
PyYAML
requests
rsa
setuptools
six
sqlparse
tzdata
unicode
Unidecode
uritemplate
urllib3